Mise à jour de 'src/app/core_logic/responsable.h'

+++ src/app/core_logic/responsable.h
@@ -1,4 +1,4 @@
 
 
 int chargementArticles(int tRef[], float tPoids[], float tVol[], float tPrix[], int tPhysique);
-void sauvegadArticles(int tRef[], float tPoids[], float tVol[], float tPrix[], float tLogique);
+void sauvegadArticles(int tRef[], float tPoids[], float tVol[], float tPrix[], int tLogique);
